added required folders
