+++ lib/logic.rb
@@ -1,0 +1,4 @@
+
+class Logic
+
+end
